moved LASUtils
