+++ src/main/java/org/terasology/lightandshadow/combat/systems/AttackSystem.java
@@ -14,7 +14,7 @@ import org.terasology.engine.entitySystem.systems.RegisterMode;
 import org.terasology.engine.entitySystem.systems.RegisterSystem;
 import org.terasology.engine.logic.characters.CharacterHeldItemComponent;
 import org.terasology.engine.logic.common.ActivateEvent;
-import org.terasology.lightandshadow.gamestate.LASUtils;
+import org.terasology.lightandshadow.LASUtils;
 import org.terasology.module.inventory.systems.InventoryManager;
 import org.terasology.module.inventory.events.DropItemRequest;
 import org.terasology.module.inventory.events.InventorySlotChangedEvent;
