+++ src/main/java/org/terasology/lightandshadow/worldgeneration/bases/BaseRasterizer.java
@@ -12,7 +12,7 @@ import org.terasology.engine.world.chunks.Chunk;
 import org.terasology.engine.world.chunks.Chunks;
 import org.terasology.engine.world.generation.Region;
 import org.terasology.engine.world.generation.WorldRasterizer;
-import org.terasology.lightandshadow.gamestate.LASUtils;
+import org.terasology.lightandshadow.LASUtils;
 
 public class BaseRasterizer implements WorldRasterizer {
     private Block blackBaseStone;
